requirements.txt
alembic>=1.4.2
certifi>=2020.4.5.1
chardet>=3.0.4
click>=7.1.1
dominate>=2.5.1
Flask>=1.1.2
Flask-Bootstrap>=3.3.7.1
Flask-Login>=0.5.0
Flask-Migrate>=2.5.3
Flask-Moment>=0.9.0
Flask-SQLAlchemy>=2.4.1
Flask-SSLify>=0.1.5
Flask-WTF>=0.14.3
gunicorn>=20.0.4
idna>=2.9
itsdangerous>=1.1.0
Jinja2>=2.11.1
Mako>=1.1.2
MarkupSafe>=1.1.1
psycopg2>=2.8.5
public>=2019.4.13
python-dateutil>=2.8.1
python-dotenv>=0.12.0
python-editor>=1.0.4
query-string>=2019.4.13
requests>=2.23.0
six>=1.14.0
SQLAlchemy>=1.3.16
urllib3>=1.25.9
visitor>=0.1.3
Werkzeug>=1.0.1
WTForms>=2.2.1
